Renamed starred to home.
